Extract chat transport handling
